Add files via upload
C++ implementation of a class that can be used to compute paths in the plane whose curvature varies linearly, quadratically or in a trapezoidal manner. The paths are computed based on explicit formulae that invoke special functions such as the Fresnel integrals and Gauss' Hypergeometric function. The paths drawn will match the endpoints of a circle with radius R subtending angle theta at its origin.